Conflicts:
	Assembly-UnityScript-vs.unityproj
	Assembly-UnityScript.unityproj
	Assets/Test Stuff/Test Scene.unity

+++ Assets/Test Stuff/InventoryManager.js
@@ -1,7 +1,7 @@
 ﻿#pragma strict
 
 
-static var pencil : float = 5;
-static var ruler : float = 1;
-static var textbook : float = 2;
-static var secret : float = 0;
+static var pencil : int = 5;
+static var ruler : int = 1;
+static var textbook : int = 2;
+static var secret : int = 0;
